// luizribeiro-angular/src/app/model/candidato.ts
export interface Candidato {
	id: Number;
	nome: String;
	cpf: String;
	rg: String;
	data_nasc: String;
	sexo: String;
	mae: String;
	pai: String;
	email: String;
	cep: String;
	endereco: String;
	numero: Number;
	bairro: String;
	cidade: String;
	estado: String;
	telefone_fixo: String;
	celular: String;
	altura: Number;
	peso: Number;
  tipo_sanguineo: String;
}